# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: StreamGate API
  description: |
    NFT-gated streaming platform API. Core flow: wallet sign-in → NFT verification → protected HLS streaming.

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## NFT Permission Control
    Content is protected by NFT ownership. Users must hold the specified NFT to access protected streaming content.
  version: 1.0.0
  contact:
    name: StreamGate Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
    description: Relative path (behind proxy)
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /nft:
    get:
      tags: [NFT]
      summary: Query NFT balance
      description: Queries wallet NFT balance for a given contract
      operationId: listNFTs
      security:
        - bearerAuth: []
      parameters:
        - name: contract
          in: query
          required: true
          schema:
            type: string
          description: NFT contract address
        - name: chain_id
          in: query
          required: true
          schema:
            type: integer
            format: int64
          description: Chain ID (e.g., 1 for Ethereum, 11155111 for Sepolia)
        - name: owner
          in: query
          schema:
            type: string
          description: Owner address (defaults to authenticated wallet)
      responses:
        "200":
          description: NFT balance info
  /nft/{id}:
    get:
      tags: [NFT]
      summary: Verify NFT token ownership
      description: Verifies ownership of a specific NFT token ID
      operationId: getNFT
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Token ID
        - name: contract
          in: query
          required: true
          schema:
            type: string
        - name: chain_id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: NFT ownership info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NFTInfo"
        "404":
          description: NFT not found
  /nft/verify:
    post:
      tags: [NFT]
      summary: Verify NFT ownership
      description: Comprehensive NFT ownership verification with caching support
      operationId: verifyNFT
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NFTVerifyRequest"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NFTVerifyResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    NFTInfo:
      type: object
      properties:
        contract_address:
          type: string
        token_id:
          type: string
        owner:
          type: string
        name:
          type: string
        symbol:
          type: string
        uri:
          type: string
    NFTVerifyRequest:
      type: object
      required: [chain_id, contract_address, token_id, owner_address]
      properties:
        chain_id:
          type: integer
          format: int64
          example: 11155111
        contract_address:
          type: string
          description: NFT contract address
        token_id:
          type: string
          description: NFT token ID
        owner_address:
          type: string
          description: Wallet address to verify ownership
    NFTVerifyResponse:
      type: object
      properties:
        verified:
          type: boolean
          description: Whether the wallet owns the NFT
        balance:
          type: integer
          format: int64
          description: Token balance (for ERC-1155)
        cached:
          type: boolean
          description: Whether result was from cache
...
security:
  - bearerAuth: []